Feat:  안 쓰는 파일 삭제, 라우트 연결
