+++ src/App.tsx
@@ -7,7 +7,7 @@ import SignUpPage from './pages/SignUpPage'
 import MainPage from './pages/MainPage'
 import NotFoundPage from './pages/NotFoundPage'
 import ToastContainer from './components/common/toast/ToastContainer'
-import TestPage from './components/find/TestPage'
+import EmailFindPage from './pages/EmailFindPage'
 
 function App() {
   const routes = useRoutes([
@@ -17,12 +17,12 @@ function App() {
       children: [
         { index: true, element: <MainPage /> },
         { path: 'mypage', element: <MyPage /> },
+        { path: '/emailfind', element: <EmailFindPage /> },
       ],
     },
     { path: '/login', element: <LoginPage /> },
     { path: '/signup', element: <SignUpPage /> },
     { path: '*', element: <NotFoundPage /> },
-    { path: '/testPage', element: <TestPage /> },
   ])
 
   return (
